refactor: split out default-app-state from store for server usage
